Integrate DataParser with engine

+++ Matcher/Trades/Trade.cs
@@ -2,12 +2,12 @@ namespace TradingEngineServer.Matcher
 {
     public class Trade
     {
-        public Trade(long price, long quantity)
+        public Trade(double price, long quantity)
         {
             Price = price;
             Quantity = quantity;
         }
-        public long Price { get; set; }
+        public double Price { get; set; }
         public long Quantity { get; set; }
 
         public String GetFormattedString()
